+++ apps/plus-task/src/tasks/release_libraries/mod.rs
@@ -1,6 +1,5 @@
 mod task;
 use task::Task;
-pub use task::{CargoToml, CargoTomlPackage};
 
 use crate::tasks::release_libraries::task::Params;
 use crate::TaskResult;
